fix helloworld java sample
