+++ helloworld/java-client/src/main/resources/META-INF/spring/service.xml
@@ -16,7 +16,7 @@
        xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
        xsi:schemaLocation="http://www.springframework.org/schema/beans http://www.springframework.org/schema/beans/spring-beans-2.5.xsd">
 
-    <bean class="com.ikurento.user.Consumer" init-method="start">
+    <bean class="org.apache.dubbo.Consumer" init-method="start">
         <!-- 声明这个类 要使用的服务名-->
         <property name="userProvider" ref="userProvider"/>
     </bean>
