+++ helloworld/java-server/src/main/resources/META-INF/spring/dubbo.provider.xml
@@ -21,14 +21,14 @@
     <!-- 应用名 -->
     <dubbo:application name="user-info-server"/>
     <!-- 连接到哪个本地注册中心 -->
-    <dubbo:registry id="ikurento" address="zookeeper://127.0.0.1:2181"/>
+    <dubbo:registry id="demoZk" address="zookeeper://127.0.0.1:2181"/>
     <!-- 用dubbo协议在20880端口暴露服务 -->
     <!-- dubbo:protocol host="127.0.0.1" / -->
     <dubbo:protocol id="dubbo" name="dubbo" host="127.0.0.1" port="20010"/>
     <!-- 声明需要暴露的服务接口 -->
-    <dubbo:service id="aaa" registry="ikurento" timeout="3000" interface="com.ikurento.user.UserProvider"
+    <dubbo:service id="aaa" registry="demoZk" timeout="3000" interface="org.apache.dubbo.UserProvider"
                    ref="demoService"/>
 
-    <bean id="demoService" class="com.ikurento.user.UserProviderImpl"/>
+    <bean id="demoService" class="org.apache.dubbo.UserProviderImpl"/>
 
 </beans>
